% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventofcode-package.R
\docType{package}
\name{adventofcode-package}
\alias{adventofcode}
\alias{adventofcode-package}
\title{adventofcode}
\description{
adventofcode solutions in R
}
\references{
\url{https://adventofcode.com/2021}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://sonia-mitchell.com/adventofcode}
}

}
\author{
\strong{Maintainer}: Sonia Mitchell \email{sonia.mitchell@glasgow.ac.uk} (\href{https://orcid.org/0000-0003-1536-2066}{ORCID})

}
\keyword{internal}
