% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-2.R
\name{check_adjacent}
\alias{check_adjacent}
\title{Return neighbouring points that have a height less than 9}
\usage{
check_adjacent(dat, row, column)
}
\description{
Return neighbouring points that have a height less than 9
}
