% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-1.R
\name{compare_heights}
\alias{compare_heights}
\title{Find `height` of adjacent points and if it's lower than all
`adjacent_points`, return risk level}
\usage{
compare_heights(dat, row, column)
}
\description{
Find `height` of adjacent points and if it's lower than all
`adjacent_points`, return risk level
}
