% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-2.R
\name{contains_subset}
\alias{contains_subset}
\title{contains_subset}
\usage{
contains_subset(unknown_vector, known_string, contains = TRUE)
}
\arguments{
\item{unknown_vector}{an unknown vector}

\item{known_string}{a known string}

\item{contains}{(optional) default is `TRUE`}
}
\description{
Find (1) which string in `unknown_vector` contains all segments
(characters) in `known_string` or (2) which string is the only one that
doesn't
}
