% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-1.R, R/10-2.R
\name{day10}
\alias{day10}
\alias{read_day10}
\alias{syntax_score}
\alias{syntax_score2}
\title{Day 10: Syntax Scoring}
\source{
<https://adventofcode.com/2021/day/10>
}
\usage{
read_day10(path)

syntax_score(dat)

syntax_score2(dat, remove_lines)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}
}
\description{
Day 10: Syntax Scoring
}
