% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-1.R, R/11-2.R, R/18-1.R, R/18-2.R, R/19-1.R,
%   R/19-2.R, R/20-1.R
\name{day11}
\alias{day11}
\alias{read_day11}
\alias{simulate_octopus}
\alias{simulate_octopus2}
\alias{read_day18}
\alias{snailfish_maths}
\alias{snailfish_maths2}
\alias{read_day19}
\alias{assemble_map}
\alias{dist_scanners}
\alias{get_algorithm}
\alias{get_input}
\alias{view_image}
\alias{enhance_image}
\title{Day 11: Dumbo Octopus}
\source{
<https://adventofcode.com/2021/day/11>

<https://adventofcode.com/2021/day/11>

<https://adventofcode.com/2021/day/11>

<https://adventofcode.com/2021/day/11>
}
\usage{
read_day11(path)

simulate_octopus(octopus)

simulate_octopus2(octopus)

read_day18(path)

snailfish_maths(dat)

snailfish_maths2(dat)

read_day19(path)

assemble_map(scans)

dist_scanners(scanners)

get_algorithm(path)

get_input(path)

view_image(img)

enhance_image(input, n, algorithm)
}
\arguments{
\item{path}{file path}

\item{octopus}{octopus}

\item{dat}{dat}

\item{scans}{scans}

\item{scanners}{scanners}

\item{img}{img}

\item{input}{input}

\item{n}{n}

\item{algorithm}{algorithm}
}
\description{
Day 11: Dumbo Octopus

Day 11: Dumbo Octopus

Day 11: Dumbo Octopus

Day 11: Dumbo Octopus
}
