% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-1.R
\name{day12}
\alias{day12}
\alias{read_day12}
\alias{plot_routes}
\title{Day 12: Passage Pathing}
\source{
<https://adventofcode.com/2021/day/12>
}
\usage{
read_day12(path)

plot_routes(dat)
}
\arguments{
\item{path}{file path}

\item{dat}{data}
}
\description{
Day 12: Passage Pathing
}
