% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-1.R
\name{day14}
\alias{day14}
\alias{get_template}
\alias{get_rules}
\alias{polymerization}
\title{Day 14: Extended Polymerization}
\source{
<https://adventofcode.com/2021/day/14>
}
\usage{
get_template(path)

get_rules(path)

polymerization(template, rules, n)
}
\arguments{
\item{path}{file path}

\item{template}{template}

\item{rules}{rules}

\item{n}{n}
}
\description{
Day 14: Extended Polymerization
}
