% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-1.R
\name{day15}
\alias{day15}
\alias{read_day15}
\alias{path_finder}
\title{Day 15: Chiton}
\source{
<https://adventofcode.com/2021/day/15>
}
\usage{
read_day15(path)

path_finder(start, end, dat)
}
\arguments{
\item{path}{file path}

\item{start}{start}

\item{end}{end}

\item{dat}{dat}
}
\description{
Day 15: Chiton
}
