% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-1.R, R/17-2.R
\name{day17}
\alias{day17}
\alias{read_day17}
\alias{simulate_launch}
\alias{move}
\alias{launch}
\alias{find_possibilities}
\alias{simulate_launch2}
\alias{possible_x}
\alias{possible_y}
\title{Day 17: Trick Shot}
\source{
<https://adventofcode.com/2021/day/17>
}
\usage{
read_day17(path)

simulate_launch(xrange, yrange)

move(status)

launch(df, trial, xrange, yrange)

find_possibilities(xrange, yrange)

simulate_launch2(velocities, xrange, yrange)

possible_x(range, min_v, max_v)

possible_y(range, min_v, max_v)
}
\arguments{
\item{path}{file path}

\item{xrange}{xrange}

\item{yrange}{yrange}

\item{status}{last step}

\item{velocities}{velocities}

\item{range}{range of target}

\item{min_v}{minimum velocity to try}

\item{max_v}{maximum velocity to try}
}
\description{
Move one step

Launch probe until target is hit or overshot

Test which velocities might hit the target

Test which velocities might hit the target
}
