% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-1.R
\name{day18}
\alias{day18}
\alias{read_day18}
\alias{snailfish_maths}
\title{Day 18: Snailfish}
\source{
<https://adventofcode.com/2021/day/18>
}
\usage{
read_day18(path)

snailfish_maths(dat)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}
}
\description{
Day 18: Snailfish
}
