% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-1.R
\name{day19}
\alias{day19}
\alias{read_day19}
\alias{assemble_map}
\title{Day 19: Beacon Scanner}
\source{
<https://adventofcode.com/2021/day/19>
}
\usage{
read_day19(path)

assemble_map(scans)
}
\arguments{
\item{path}{file path}

\item{scans}{scans}
}
\description{
Day 19: Beacon Scanner
}
