% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1.R, R/2-2.R
\name{day2}
\alias{day2}
\alias{read_day2}
\alias{dive}
\alias{dive2}
\title{Day 2: Dive!}
\source{
<https://adventofcode.com/2021/day/2>
}
\usage{
read_day2(path)

dive(dat)

dive2(dat)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}
}
\description{
Day 2: Dive!
}
