% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-1.R
\name{day20}
\alias{day20}
\alias{get_algorithm}
\alias{get_input}
\alias{view_image}
\alias{enhance_image}
\title{Day 20: Trench Map}
\source{
<https://adventofcode.com/2021/day/20>
}
\usage{
get_algorithm(path)

get_input(path)

view_image(img)

enhance_image(input, n, algorithm)
}
\arguments{
\item{path}{file path}

\item{img}{img}

\item{input}{input}

\item{n}{n}

\item{algorithm}{algorithm}
}
\description{
Day 20: Trench Map
}
