% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21-1.R
\name{day21}
\alias{day21}
\alias{read_day21}
\alias{deterministic}
\title{Day 21: Dirac Dice}
\source{
<https://adventofcode.com/2021/day/21>
}
\usage{
read_day21(path)

deterministic(dat)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}
}
\description{
Day 21: Dirac Dice
}
