% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22-1.R
\name{day22}
\alias{day22}
\alias{read_day22}
\alias{reboot}
\title{Day 22: Reactor Reboot}
\source{
<https://adventofcode.com/2021/day/22>
}
\usage{
read_day22(dat, limit)

reboot(dat)
}
\arguments{
\item{dat}{dat}

\item{path}{file path}
}
\description{
Day 22: Reactor Reboot
}
