% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-1.R
\name{day24}
\alias{day24}
\alias{read_day24}
\title{Day 24: Arithmetic Logic Unit}
\source{
<https://adventofcode.com/2021/day/24>
}
\usage{
read_day24(path)
}
\arguments{
\item{path}{file path}
}
\description{
Day 24: Arithmetic Logic Unit
}
