% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25-1.R
\name{day25}
\alias{day25}
\alias{read_day25}
\alias{get_cucumbers}
\alias{simulate_cucumbers}
\title{Day 25: Sea Cucumber}
\source{
<https://adventofcode.com/2021/day/25>
}
\usage{
read_day25(path)

get_cucumbers(dat)

simulate_cucumbers(dat, cucumbers)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}

\item{cucumbers}{cucumbers}
}
\description{
Read in cucumber positions
}
