% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1.R, R/3-2.R
\name{day3}
\alias{day3}
\alias{read_day3}
\alias{power_consumption}
\alias{rating}
\title{Day 3: Binary Diagnostic}
\source{
<https://adventofcode.com/2021/day/3>
}
\usage{
read_day3(path)

power_consumption(dat)

rating(dat, type)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}

\item{type}{type}
}
\description{
Day 3: Binary Diagnostic
}
