% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-1.R, R/4-2.R
\name{day4}
\alias{day4}
\alias{get_numbers}
\alias{get_boards}
\alias{play_bingo}
\alias{lose_bingo}
\title{Day 4: Giant Squid}
\source{
<https://adventofcode.com/2021/day/4>
}
\usage{
get_numbers(path)

get_boards(path)

play_bingo(numbers, boards)

lose_bingo(numbers, boards)
}
\arguments{
\item{path}{file path}

\item{numbers}{numbers}

\item{boards}{boards}
}
\description{
Day 4: Giant Squid
}
