% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-1.R, R/5-2.R
\name{day5}
\alias{day5}
\alias{read_day5}
\alias{track_vents}
\alias{track_vents2}
\title{Day 5: Hydrothermal Venture}
\source{
<https://adventofcode.com/2021/day/5>
}
\usage{
read_day5(path)

track_vents(coordinates)

track_vents2(coordinates)
}
\arguments{
\item{path}{file path}

\item{coordinates}{coordinates}
}
\description{
Day 5: Hydrothermal Venture
}
