% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-1.R, R/6-2.R
\name{day6}
\alias{day6}
\alias{read_day6}
\alias{simulate_lanternfish}
\alias{simulate_lanternfish2}
\title{Day 6: Lanternfish}
\source{
<https://adventofcode.com/2021/day/6>
}
\usage{
read_day6(path)

simulate_lanternfish(dat)

simulate_lanternfish2(dat)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}
}
\description{
Day 6: Lanternfish
}
