% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-1.R, R/7-2.R
\name{day7}
\alias{day7}
\alias{read_day7}
\alias{track_crabs}
\alias{track_crabs2}
\title{Day 7: The Treachery of Whales}
\source{
<https://adventofcode.com/2021/day/7>
}
\usage{
read_day7(path)

track_crabs(dat)

track_crabs2(dat)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}
}
\description{
Day 7: The Treachery of Whales
}
