% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-1.R, R/8-2.R
\name{day8}
\alias{day8}
\alias{read_day8}
\alias{count_digits}
\alias{decode_segments}
\title{Day 8: Seven Segment Search}
\source{
<https://adventofcode.com/2021/day/8>
}
\usage{
read_day8(path)

count_digits(dat)

decode_segments(dat)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}
}
\description{
Day 8: Seven Segment Search
}
