% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-1.R, R/9-2.R
\name{day9}
\alias{day9}
\alias{read_day9}
\alias{low_points}
\alias{survey_basins}
\title{Day 9: Smoke Basin}
\source{
<https://adventofcode.com/2021/day/9>
}
\usage{
read_day9(path)

low_points(dat)

survey_basins(dat, lowest_points)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}

\item{lowest_points}{lowest_points}
}
\description{
Day 9: Smoke Basin
}
