% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-2.R
\name{find_digit}
\alias{find_digit}
\title{Find string in `string_vector` of length `num_characters`}
\usage{
find_digit(string_vector, num_characters)
}
\description{
Find string in `string_vector` of length `num_characters`
}
