% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-2.R
\name{subset_of}
\alias{subset_of}
\title{Find the string in `unknown_vector` whose segments (characters) are a
subset of `known_string`}
\usage{
subset_of(unknown_vector, known_string)
}
\description{
Find the string in `unknown_vector` whose segments (characters) are a
subset of `known_string`
}
