% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-2.R
\name{survey}
\alias{survey}
\title{For a particular `lowest_point`, survey the basin and return it's size}
\usage{
survey(dat, row, column)
}
\description{
For a particular `lowest_point`, survey the basin and return it's size
}
