% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-1.R, R/11-2.R, R/18-2.R, R/19-2.R
\name{day11}
\alias{day11}
\alias{read_day11}
\alias{simulate_octopus}
\alias{simulate_octopus2}
\alias{snailfish_maths2}
\alias{dist_scanners}
\title{Day 11: Dumbo Octopus}
\source{
<https://adventofcode.com/2021/day/11>
}
\usage{
read_day11(path)

simulate_octopus(octopus)

simulate_octopus2(octopus)

snailfish_maths2(dat)

dist_scanners(scanners)
}
\arguments{
\item{path}{file path}

\item{octopus}{octopus}

\item{dat}{dat}

\item{scanners}{scanners}
}
\description{
Day 11: Dumbo Octopus
}
