% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-1.R
\name{day13}
\alias{day13}
\alias{read_day13}
\alias{origami_instructions}
\alias{origami_paper}
\alias{fold_left}
\title{Day 13: Transparent Origami}
\source{
<https://adventofcode.com/2021/day/13>
}
\usage{
read_day13(path)

origami_instructions(dat)

origami_paper(dat)

fold_left(grid, value)
}
\arguments{
\item{path}{file path}

\item{dat}{dat}

\item{grid}{grid}

\item{value}{value}
}
\description{
Day 13: Transparent Origami
}
