% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-1.R
\name{day17}
\alias{day17}
\alias{read_day17}
\alias{simulate_launch}
\title{Day 17: Trick Shot}
\source{
<https://adventofcode.com/2021/day/17>
}
\usage{
read_day17(path)

simulate_launch(xrange, yrange)
}
\arguments{
\item{path}{file path}

\item{xrange}{xrange}

\item{yrange}{yrange}
}
\description{
Day 17: Trick Shot
}
